requirements-dev.txt
# Core dependencies (mirror from requirements.txt)
-r requirements.txt

# Development tools
django-debug-toolbar
ipython
python-dotenv

# Code quality
black
flake8
isort
mypy

# Testing
pytest
pytest-django
pytest-cov
coverage

# Documentation
sphinx
sphinx-rtd-theme

# Debugging
django-extensions
werkzeug

# Performance testing
locust
